Create a view controller to manage routings of views
